propiedades/src/modulos/propiedad/infraestructura/schema/v1/eventos.py
from pulsar.schema import *


class TransaccionCreadaPayload():
    id_propiedad = String()
    compania_destino_id = String()


class DatosGeograficosActualizadosPayload():
    id_propiedad = String()
    latitud = Float()
    longitud = Float()
